+++ src/usefulgnom/serialize/__init__.py
@@ -1,9 +1,8 @@
 """Handel serialization of objects to and from strings."""
 
-from usefulgnom.serialize.basecnt_coverage import load_convert_bnc
-from usefulgnom.serialize.total_coverage import load_convert_total
 
 all = [
     "load_convert_bnc",
     "load_convert_total",
-    ]
+    "extract_sample_ID",
+]
